% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_outputs_backup.R
\name{create_outputs_backup}
\alias{create_outputs_backup}
\title{Move Figures and tables to google drive}
\usage{
create_outputs_backup(tar_obj, ..., google_drive_folder = "Figures and Tables")
}
\arguments{
\item{tar_obj}{Target. A target object to be backed up the output of tar_file}

\item{...}{additional targets to be backed up}

\item{google_drive_folder}{String. Where should files go in Google drive}
}
\value{
Documentation of upload
}
\description{
Move Figures and tables to google drive
}
