% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{dms_to_deg}
\alias{dms_to_deg}
\title{Convert dms to degree}
\usage{
dms_to_deg(x)
}
\arguments{
\item{x}{(d)egree (m)inute (s)econd vector of locations}
}
\value{
degree locations
}
\description{
Convert dms to degree
}
