% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/establish_calendar.R
\name{establish_calendar}
\alias{establish_calendar}
\title{Set up a calendar for Matlapitsi Rousettus for use in figures.}
\usage{
establish_calendar()
}
\value{
Data frame with date identifiers for figure creation to highlight specific timing of Rousettus life history at Matlapitsi
}
\description{
Set up a calendar for Matlapitsi Rousettus for use in figures.
}
