% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_file_paths.R
\name{generate_file_paths_linked}
\alias{generate_file_paths_linked}
\title{Build some file paths for saving outputs}
\usage{
generate_file_paths_linked(tar_obj, pattern, ...)
}
\arguments{
\item{tar_obj}{Saved tar object built from a tar_make call}

\item{pattern}{file extension}

\item{...}{additional arguments}
}
\value{
Local file paths for saving objects
}
\description{
Build some file paths for saving outputs
}
