% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_base_id}
\alias{get_base_id}
\title{Get base ID from Sys.getenv() to work in Actions workflow}
\usage{
get_base_id()
}
\value{
Airtable base IDs from .yml / github actions
}
\description{
Get base ID from Sys.getenv() to work in Actions workflow
}
