% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sequence_data.R
\name{get_each_sequence}
\alias{get_each_sequence}
\title{Guts of the internal lapply within get_sequence_data. This function not to be run separately, run the wrapper get_sequence_data}
\usage{
get_each_sequence(x, dir_name)
}
\arguments{
\item{x}{a sequence file}

\item{dir_name}{name of local directory where sequence files will temporarily get stored for parsing}
}
\value{
a data frame with a stripped sequence (character string) and identifier information for this sample
}
\description{
Guts of the internal lapply within get_sequence_data. This function not to be run separately, run the wrapper get_sequence_data
}
