% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GPS_data.R
\name{get_GPS_data_airtable}
\alias{get_GPS_data_airtable}
\title{Download GPS files from Airtable Base}
\usage{
get_GPS_data_airtable(
  base = "appi9YHtonEpMFuAz",
  table_name = "GPS Data",
  field = "GPS Data Processed",
  dir_name = "data/temp_raw_gps_data"
)
}
\arguments{
\item{base}{String. Base ID}

\item{table_name}{String. Table nanme}

\item{field}{List. Fields to be included in extract}

\item{dir_name}{Location to temporarily store sequence file for processing}
}
\value{
data frame of stacked GPS tracks
}
\description{
Download GPS files from Airtable Base
}
