% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sabrenet_data.R
\name{get_sabrenet_data}
\alias{get_sabrenet_data}
\title{pull data down from airtable}
\usage{
get_sabrenet_data(
  base = list("appi9YHtonEpMFuAz"),
  table_name = "Sample Data",
  view = "For Export",
  fields,
  ...
)
}
\arguments{
\item{base}{Airtable base}

\item{table_name}{Airtable table name}

\item{view}{View for the Airtable table name desired}

\item{fields}{which fields to download}

\item{...}{anything else}
}
\value{
A data frame of data from Airtable
}
\description{
pull data down from airtable
}
