% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sequence_data.R
\name{get_sequence_data}
\alias{get_sequence_data}
\title{Get Sequence Data -- Extract sequence data from Specimen Results and strip the sequence from the uploaded .qb or .fas files}
\usage{
get_sequence_data(
  specimen_data = specimen_data,
  cleaned_sequences,
  base = base_id,
  dir_name = "data/sequence_data",
  field = "Consensus Sequence"
)
}
\arguments{
\item{specimen_data}{Specimen data}

\item{cleaned_sequences}{Sequences from Sequences table in Airtable (already stripped from file attachments in Specimen Data)}

\item{base}{base id}

\item{dir_name}{Location to temporarily store sequence file for processing}

\item{field}{string to look for for sequence attachments}
}
\value{
Data frame of sequences for all entries of sequence data that have files uploaded
}
\description{
Get Sequence Data -- Extract sequence data from Specimen Results and strip the sequence from the uploaded .qb or .fas files
}
