% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_GPS_data.R
\name{process_GPS_data}
\alias{process_GPS_data}
\title{Process GPS Data -- Clean raw ingested GPS tracks from all bats}
\usage{
process_GPS_data(
  GPS_data,
  outlier_pos,
  outlier_date,
  ping_time_test,
  file_type,
  epd
)
}
\arguments{
\item{GPS_data}{Ingested GPS data}

\item{outlier_pos}{Obvious outlier GPS data (previously hand-built data frame of min and max for long and lat (two columns, two rows))}

\item{outlier_date}{Drop data points before some date?}

\item{ping_time_test}{Duty cycle (in min) that we want to test. NA to ignore this argument}

\item{file_type}{Whether GPS was ingested via kml or txt}

\item{epd}{Errant Point Directory. Location where a file exists that lists additional points to be stripped from the GPS data}
}
\value{
list of data frame of cleaned GPS data and ping time
}
\description{
Process GPS Data -- Clean raw ingested GPS tracks from all bats
}
