% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_GPS_data.R
\name{process_GPS_data_for_plotting}
\alias{process_GPS_data_for_plotting}
\title{Process GPS Data for Plotting -- Further cleaning and creating of additional fields for plotting}
\usage{
process_GPS_data_for_plotting(
  GPS_data,
  bin_dist,
  min_flight_dist,
  time_window,
  ping_time
)
}
\arguments{
\item{GPS_data}{Data frame of cleaned GPS data}

\item{bin_dist}{Width in c(long, lat) to join nearby datapoints to count as time spent in the same spot}

\item{min_flight_dist}{Distance traveled between adjacent data points at which 'flight' is assigned}

\item{time_window}{Number of points on either side of a focal point not identified as 'flight' to characterize that point as 'feeding'}

\item{ping_time}{Number of minutes that elapse between GPS records}
}
\value{
data frame of stacked GPS tracks
}
\description{
Process GPS Data for Plotting -- Further cleaning and creating of additional fields for plotting
}
