% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_virus_data.R
\name{process_summarized_sampling_data_for_plotting}
\alias{process_summarized_sampling_data_for_plotting}
\title{Process Summarized Sampling Data for Plotting}
\usage{
process_summarized_sampling_data_for_plotting(
  sample_data,
  wanted_site,
  name_for_sampling_plot
)
}
\arguments{
\item{sample_data}{Data frame. Already summarized sample data}

\item{wanted_site}{Desired site to isolate for further summary}

\item{name_for_sampling_plot}{Common name for plot of sampling over time}
}
\value{
Data frame (further summarized sampling data)
}
\description{
Process Summarized Sampling Data for Plotting
}
