% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_virus_data.R
\name{process_summarized_specimen_data_for_plotting}
\alias{process_summarized_specimen_data_for_plotting}
\title{Process Summarized Specimen Data for Plotting}
\usage{
process_summarized_specimen_data_for_plotting(
  specimen_data,
  grouping_vec,
  wanted_site,
  sample_type
)
}
\arguments{
\item{specimen_data}{Data frame. Already summarized specimen data}

\item{grouping_vec}{Character vector. Variables to group over to summarize}

\item{wanted_site}{Desired site to isolate for further summary}

\item{sample_type}{individual or environmental}
}
\value{
Data frame (further summarized specimen data)
}
\description{
Process Summarized Specimen Data for Plotting
}
