% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sequence_data.R
\name{read.gb}
\alias{read.gb}
\title{Function to extract sequence from .gb files. Called from within get_each_sequence}
\usage{
read.gb(z)
}
\arguments{
\item{z}{a .gb file}
}
\value{
a stripped sequence (character string)
}
\description{
Function to extract sequence from .gb files. Called from within get_each_sequence
}
