% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sabrenetfunctions-package.R
\docType{package}
\name{sabrenetfunctions-package}
\alias{sabrenetfunctions}
\alias{sabrenetfunctions-package}
\title{sabrenetfunctions: SABRENet Functions}
\description{
Functions for downloading, modifying, summarizing, and backing up outputs from summarizing SABRENet and BEEZ data (which is housed in airtable).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ecohealthalliance/sabrenetfunctions}
}

}
\author{
\strong{Maintainer}: Morgan Kain \email{kain@ecohealthalliance.org} (\href{https://orcid.org/0000-0003-0605-7289}{ORCID})

}
\keyword{internal}
