% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_tables.R
\name{save_tables}
\alias{save_tables}
\title{Save ethics reporting table and other DTRA reporting tables locally}
\usage{
save_tables(ethics_table, DTRA_tables)
}
\arguments{
\item{ethics_table}{Data frame. Ethics table}

\item{DTRA_tables}{List of data frames (DTRA tables).}
}
\value{
Data frame. Summary of sampling by Year, Month, Site
}
\description{
Save ethics reporting table and other DTRA reporting tables locally
}
