% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{split_tibble}
\alias{split_tibble}
\title{Split up data frame into list with entries defined by 'col'}
\usage{
split_tibble(tibble, col = "col")
}
\arguments{
\item{tibble}{Tibble for splitting into list}

\item{col}{single or multiple columns over which the tibble will be split up into list entries}
}
\value{
List of tibbles split according to the argument col
}
\description{
Split up data frame into list with entries defined by 'col'
}
