% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_virus_data.R
\name{summarize_virus_data}
\alias{summarize_virus_data}
\title{Summarize Virus Data}
\usage{
summarize_virus_data(
  bat_data,
  site_data,
  specimen_data,
  sequence_data,
  base_ids
)
}
\arguments{
\item{bat_data}{Data frame. Sample data}

\item{site_data}{Data frame. Site info}

\item{specimen_data}{Data frame. Specimen data}

\item{sequence_data}{Data frame. Sequences (all the already cleaned ones in the Airtable table "Sequences")}

\item{base_ids}{list of two Airtable base IDs}
}
\value{
List of two data frames (specimen data and sample data)
}
\description{
Summarize Virus Data
}
